backend/requirements.txt
# Core LLM / RAG framework
langchain==0.3.25
langchain-openai==0.3.8
langchain-community==0.3.23
langchain-text-splitters==0.3.8

# Vector database
faiss-cpu==1.10.0

# Document loading
pypdf==5.4.0

# OpenAI SDK (required by langchain-openai)
openai==1.66.3

# Web framework
flask==3.1.0
flask-cors==5.0.1

# Utilities
python-dotenv==1.0.1
werkzeug==3.1.3
tiktoken==0.9.0

